+++ src/main/java/br/com/saude/models/WeightLossRecord.java
@@ -1,7 +1,5 @@
 package br.com.saude.models;
 
-import java.math.BigDecimal;
-
 import javax.persistence.*;
 
 import lombok.Getter;
@@ -14,5 +12,5 @@ import lombok.Setter;
 public class WeightLossRecord extends Record {
 
     @Column(name="current_weight", nullable=false)
-    private BigDecimal currentWeight;
+    private int currentWeight;
 }
